fix: resolve PostgreSQL type casting issues and improve multi-cloud support
## Major Fixes
- **PostgreSQL JOIN compatibility**: Add ::INTEGER type casting in folders.js for created_by column (5 locations)
- **Database schema**: Add missing parent_id column to PostgreSQL folders table
- **Clean slate migration**: Recreate folders table with proper integer types and default data
- **Authentication flow**: Fix boolean parameter handling across auth, users, and metadata routes
- **Deploy script**: Correct region from us-central1-b to us-central1

## Schema & Database Improvements
- **Multi-database support**: Enhanced SQLite/PostgreSQL compatibility
- **Schema migrations**: Add automatic column type conversion and data migration
- **Error handling**: Improve database connection resilience in production

## Security Enhancements
- **Environment management**: Add .env.example templates for backend and frontend
- **Secrets protection**: Update .gitignore to exclude all sensitive files (.env, .secret, .key)
- **Documentation**: Add security configuration section to CLAUDE.md

## Configuration Updates
- **GCP Cloud Run**: Update API URLs to correct format (1044697249626.us-central1.run.app)
- **CORS setup**: Support both local and production frontend domains
- **OAuth setup**: Document proper Google OAuth client configuration

## Testing Status
- ✅ Folders now load correctly in GCP environment after login
- ✅ Google OAuth authentication working
- ✅ PostgreSQL type compatibility resolved
- ✅ Local SQLite environment unaffected

+++ backend/src/routes/folders.js
@@ -17,7 +17,7 @@ router.get('/', async (req, res, next) => {
              (SELECT COUNT(*) FROM certificates WHERE folder_id = f.id) as certificate_count,
              pf.name as parent_name
       FROM folders f
-      LEFT JOIN users u ON f.created_by = u.id
+      LEFT JOIN users u ON f.created_by::INTEGER = u.id
       LEFT JOIN folders pf ON f.parent_id = pf.id
       WHERE 1=1
     `;
@@ -48,7 +48,7 @@ router.get('/:id', validateId, async (req, res, next) => {
              (SELECT COUNT(*) FROM certificates WHERE folder_id = f.id) as certificate_count,
              pf.name as parent_name
       FROM folders f
-      LEFT JOIN users u ON f.created_by = u.id
+      LEFT JOIN users u ON f.created_by::INTEGER = u.id
       LEFT JOIN folders pf ON f.parent_id = pf.id
       WHERE f.id = ?
     `, [id]);
@@ -117,7 +117,7 @@ router.post('/', validateFolder, requirePermission('folders:write'), async (req,
     const folder = await db.getAsync(`
       SELECT f.*, u.username as created_by_username
       FROM folders f
-      LEFT JOIN users u ON f.created_by = u.id
+      LEFT JOIN users u ON f.created_by::INTEGER = u.id
       WHERE f.id = ?
     `, [folderId]);
 
@@ -163,7 +163,7 @@ router.put('/:id', validateId, validateFolder, requirePermission('folders:write'
     const updatedFolder = await db.getAsync(`
       SELECT f.*, u.username as created_by_username
       FROM folders f
-      LEFT JOIN users u ON f.created_by = u.id
+      LEFT JOIN users u ON f.created_by::INTEGER = u.id
       WHERE f.id = ?
     `, [id]);
 
@@ -299,7 +299,7 @@ router.patch('/:id/move', validateId, requirePermission('folders:write'), async 
              (SELECT COUNT(*) FROM certificates WHERE folder_id = f.id) as certificate_count,
              pf.name as parent_name
       FROM folders f
-      LEFT JOIN users u ON f.created_by = u.id
+      LEFT JOIN users u ON f.created_by::INTEGER = u.id
       LEFT JOIN folders pf ON f.parent_id = pf.id
       WHERE f.id = ?
     `, [id]);
